Update patch set 5
Patch Set 5:

b/112189958 says the debugfs work is being punted to R. Although we're adding one more dependency, it's one more amongst many.

If we really can't do this, I'm not familiar with alternative options (I am responding on behalf of rslawik@ who is OOO.) Can you provide any guidance?

Background: this access will power a memory impact metric (total ion consumption) that will be very useful for debugging low memory conditions in media apps (in particular we are keen on getting data for Camera.) This is a late breaking requirement and we're trying to get this into Q before the final deadlines.

Patch-set: 5