Update patch set 1
Patch Set 1:

This change makes me nervous, because it introduces a large difference in behavior between userdebug and user builds. The userdebug_or_eng() macro is dangerous because of the differences it introduces between build types. Most testing is performed on userdebug builds, not user builds. The lack of permissions on user builds could mean that bugs remain latent until it's too late. This is especially bad for fastbootd, which is needed for flashing the device and other OTA-ish activity. A bug here could be very bad.

I don't know what the right solution is. Can we minimize the differences between build types to reduce the probability of bugs?

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>