Update patch set 1
Patch Set 1:

NB: A properly functioning adb remount is blocked by this, I am worried that chatty is going to start complaining (it already has been noticed, but the impact has yet to be digested). The impact is that if 'vendor' is flashed, the overrides for 'vendor' still exist and are not wiped, which could result in confusing effects on development. This is also blocking aosp/930316 which requires:

    allow fastbootd proc_filesystems:file { getattr open read };
    allow fastbootd self:capability sys_rawio;

to perform the filesystem probes.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0