Update patch set 1
Patch Set 1: Autosubmit+1

Summary:

We are, and need to continue, to followup on nnk@'s concerns.

The current inability to capture the kernel logs from a fastbootd session prevents us from instrumenting avc problems b/128687557. Even if propagated, there is currently no infrastructure to monitor it. Whatever fixes to b/128687557, or existing (?) recovery logs, needs to be part of the security teams avc capture. This vacuum needs to be filled.

Functional tests might not be enough, for instance fastbootd continues to function even though it is currently and recently failing to wipe a partition's overrides (adb-reboot-test.sh catches it, but also can not be run automatically because of b/117670584). This behavior regressed last week and only a keen observation of someone complaining about the noise in his console logs while fastbootd was demonstrating a bug caught it before it became serious once whatever tightening regressed it merged down and built from hourly ToT sync.

adb-reboot-test.sh might have to be run semi-manually in Q/A, will look into that.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=+1
Label: Presubmit-Ready=0