Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> If we should move the dump into device specific hals, it means there will have two different mechanism for userdebug and user build. My thinking is that mechanism should keep the same, using ::debug method no matter userdebug or user build. That is the reason to grant permission for dumpstate. In this change, we grant the permission to dumpstate and IWifi::debug is able be called by dumpstate in user build.

There isn't a different mechanism on userdebug and user builds, you're running dumpstate as root. It's WAI that su is unrestricted on debug builds, but that's not a good argument that it should work the same way on a user build.

Steven, I'm asking if this is acceptable from a Treble compliance standpoint. The neverallow rule being changed was intended to enforce Treble.

Patch-set: 1