Update patch set 1
Patch Set 1:

(18 comments)

Another way to handle this would be to introduce a new attribute to cover both, mark it expandattribute, and then rephrase all the rules to use it.
But that's problematic for neverallow rules, and isn't obviously better for the end goal of minimising the rules that refer to the new type.

(In a couple of cases the way forward is to remove the old rule for system_data_file and only leave the new rule, of course.)

Patch-set: 1
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1060831 <1060831@85c56323-6fa9-3386-8a01-6480fb634889>