Update patch set 6
Patch Set 6:

Hi, as Yongqin mentions it looks like there is a problem with webview here; have you tested that this works with chrome/webview at the version that's currently included in AOSP master (66)? Or only with the latest version on the Play Store/in internal master?

It's possible that the dependency on /dev/ashmem has been removed in newer versions, but we don't normally update the version in AOSP outside of a platform release unless there's a specific problem with it, so if that's the case we need to know so we can update it..

Patch-set: 6
CC: Gerrit User 1021004 <1021004@85c56323-6fa9-3386-8a01-6480fb634889>