Update patch set 6
Patch Set 6:

Oh - further issue here is that since WebView is loaded directly into apps, this policy will be enforced if the *app* targets Q, even if the build of WebView the app is using does not. This won't be a problem once Q is released since the release version will have to contain a WebView that does target Q and is compatible, but master currently doesn't.

Patch-set: 6