Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Do you have a bug for this?
> When did traced_probes start talking to atrace?

b/127378737 has some details but note that we changed from the original approach.
We are not currently talking to the atrace HAL but I have a CL that is WIP to do just that.
I can upload it later today if you'd prefer seeing that before +2ing this.

Patch-set: 1
CC: Gerrit User 1040443 <1040443@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1269255 <1269255@85c56323-6fa9-3386-8a01-6480fb634889>