Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Tri, Jiyong, can you all comment on this? Looks like aosp/930776 has the same issue. Seems like the simplest solution is to set this in the system build.prop instead of the vendor one.

I agree with you. I didn't notice that the bluetooth app has write access to the sysprop. In fact, all bluetooth_*_prop seem to have the same problem.

Ray, Cheney, I think access to these sysprops should be hidden behind the HIDL interface or be schematized using the sysprop-as-api (go/android_sysprop_as_api). The bottom line is that any information that across the system and vendor partition should be explicitly specified as an interface and the interface should be kept in a backward compatible manner.

Patch-set: 5
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>