Update patch set 3
Patch Set 3: Code-Review-1

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Are there actually APIs that we expose to developers, or is this solely exposed for CTS testing purposes? I'm worried about exposing unnecessary attack surface to applications.
> 
> At this point, there are no "public" APIs. They're all @SystemApi.

Annotating the service definition with "app_api_service" makes this service available to all Android applications. It becomes a de-facto API, even if the intention is that it's not an API. If there are no public APIs which are intended to use this functionality, then exposing this attack surface to applications seems improper.

Patch-set: 3
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1