Update patch set 12
Patch Set 12:

(1 comment)

> Patch Set 12:
> 
> additionally, if this is an architectural property you want to maintain, you should add neverallow rules, to prevent regressions. Otherwise someone may add further allow rules, either in core policy or device specific policy.

Thanks for the suggestion. I agree. For now, I've create a neverallow rule only for appdomain. I will try to harden it (e.g. domain except for init, ueventd, and apexd) in a follow-up CL as I am not sure whether or not it cause any trouble in downstream.

Patch-set: 12
Label: Presubmit-Ready=0