Update patch set 1
Patch Set 1:

> Sure, but the bug is not prevented by preventing the SIGCHLD. Programs are supposed to send their parent a SIGCHLD on exit.

Yes, I agree that the bug isn't prevented by us not handling sigchld. Programs are suppose to send their parent a sigchld on exit, but these programs shouldn't be exiting. The SELinux denial is the canary in the coal mine letting us know something is behaving badly. We (Android, not you and I) can choose to kill the canary, or fix the underlying issue causing the canary to die.

On the other hand, we can also acknowledge that crashing is a common part of these programs, and so we should just accept reality. 

I'm supportive of this change, but I also think we (Android) should be following up on why these crashes are even occurring in the first place.

Still +2d

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>