Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review+2
> 
> I don't have a problem with this specific change.
> But it sounds slightly curious that one producer can finalize (or do anything) with a trace started by a different producer. Does traced implement some ACL?

Very good question. Here's how it works:

Until now (before b/130135730):
- a consumer configure, starts and consumes the trace
- a producer can only inject protobuf data into the trace.

As part of b/130135730:
- A producer can also send a trigger (a string) which says "hey traced, in case you are interested, X just happened". The specific cases are "hey I detected a slow transition in my app" or "hey, my task queues are taking forever to execute".
- A consumer can configure,start a trace and further specify "this trace should be finalized if any producer sends a trigger named "X" (producers can lie about that but it's not a big deal, they would, at worst, cause traces that are sensitive to "X" to end prematurely).

So, the ability of producers to send triggers itself is harmless. It has a defined semantic only if the trace started by consumer expects it.

This CL:
In some cases we want the cmdline itself to send the trigger (use case: the trigger comes from the platform, which knows only how to invoke the cmdline client, not from an app), hence we make it behave like a producer to send the trigger.

Patch-set: 7
Reviewer: Gerrit User 1040443 <1040443@85c56323-6fa9-3386-8a01-6480fb634889>