Allow idmap1 to read vmdl*.tmp APK install files
When upgrading a package, PackageParser acts on the temporary
APK file copied from the install location. This is passed to
idmap, which doesn't have read access because it's missing an
SELinux rule.

This is needed to fix a bug with manifest overlaying on updating
an app, a feature kept alive for Q.

Relevant logs when updating a target:
2019-04-24 09:41:17.920 6847-6847/? D/idmap: error: failed to read apk /data/app/vmdl304236685.tmp/base.apk: Permission denied
2019-04-24 09:41:17.903 6847-6847/? W/idmap: type=1400 audit(0.0:3784): avc: denied { search } for name="vmdl304236685.tmp" dev="vdc" ino=8769 scontext=u:r:idmap:s0 tcontext=u:object_r:apk_tmp_file:s0 tclass=dir permissive=0
2019-04-24 09:41:17.923 6848-6848/? W/idmap: type=1400 audit(0.0:3785): avc: denied { search } for name="vmdl304236685.tmp" dev="vdc" ino=8769 scontext=u:r:idmap:s0 tcontext=u:object_r:apk_tmp_file:s0 tclass=dir permissive=0
2019-04-24 09:41:17.940 6848-6848/? D/idmap: error: failed to read apk /data/app/vmdl304236685.tmp/base.apk: Permission denied

Bug: 130559507

Test: manual adb push /system/product/app/TestApp.apk with
/system/product/overlay/TestOverlay.apk enabling disabled launcher
Activity in TestApp; adb install -r TestApp.apk keeps enabled state
with changes

Change-Id: Ieeb7fb4f79ae091d0febf42ca358e7ffdfa6c3ff