Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)

Perhaps this stems from the fact that system_server hosts a lot of services? Is it possible to figure out what service in system_server uses mediaserver_service? If it's not audio_service then the observed circular dependency is due to the fact that there are two services in the same host.

Patch-set: 2
Reviewer: Gerrit User 1042638 <1042638@85c56323-6fa9-3386-8a01-6480fb634889>