Update patch set 2
Patch Set 2:

This looks too broad. The property type is world-readable. If the intention is to let dumpstate to read it, I think you'd better to create a new type (or find an existing one) with less accessibility.

Also, I wonder if the dumpstate process is a system one or a vendor one. If it's a system one, you don't need to list the property here; this is for sysprops exported from system to vendor. Listing here means you need to keep the property for several releases due to backwards compatibility. Is that what you mean?

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>