Update patch set 1
Patch Set 1: Code-Review+1

LGTM, but are we still in time to get this into Q? Or is Q sepolicy frozen? If we can't get this into Q, then we may be probably better off not submitting the other changes into AOSP either in order. Otherwise we'll increase divergence between Q and AOSP and that will make it more difficult to fix the IPv6-only tethering bug in Q.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1