Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The rules for nativezygote and nativezygote_vendor are identical and could be merged. You can still have two binaries, but run each of the binaries under the nativezygote domain. Is there a reason why you want them separate?

I want them to be separate so that we can specify which one to use for which service in sepolicy. The alternative would be to have a single selinux domain for both and have init decide which one to use. If we go down that route, we'd have to make sure init always make the right decision w.r.t. linker config. The way it is now, we still have to make sure we use the right native zygote instance, but that'd be a manual process and we don't have to add complicated logic in init.

Patch-set: 1
Reviewer: Gerrit User 1349653 <1349653@85c56323-6fa9-3386-8a01-6480fb634889>