Update patch set 2
Patch Set 2: Code-Review+1

> LGTM, but please wait for agampe to review. It sounds like there's a debate whether the symlink should be there at all (b/133864678) ...

I am OK with this. I just think that since we force a single version (and for having consistency) it seems wrong to use symlinks. That might be a restriction of the build system, though (we might need the symlink for the host for consistency in the host apex?).

Patch-set: 2
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1