Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can we modify this so that zygote creates the page, then the child process mprotect()s the page executable on zygote specialization? This would help drastically in ensuring that zygote (a very powerful process) can only load code which comes from dm-verity.

If I understand you correctly, that would mean zygote could not execute the generated code? That would be a problem as we'd want to run optimized code for the zygote. It also poses the problem of every child zygote to update the entrypoints of methods, which is costly in memory.

Note that the only input for the zygote JIT is files from dm-verity partitions.

Patch-set: 1
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>