Update patch set 1
Patch Set 1:

This doesn't make a lot of sense to me. It appears, from reading the bug, that the unprivileged shell user is attempting to manipulate dm devices, which it shouldn't be able to do. In particular, the unprivileged shell user cannot open /dev/loop-control, run the LOOP_* ioctls, etc. It doesn't make a lot of sense to me that apps should be able to provide memfd file descriptors which are eventually used by the kernel for dm related functionality.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>