Update patch set 3
Patch Set 3:

The bug is talking about dividing the 'oem' aka 'vendor' range of AIDs into separate ranges per partition, such that we don't have any conflicts if the system, vendor, odm, product, etc partitions are updated independently of each other (also if the bionic APEX gets updated independently of a vendor-modified system image).

To still be able to map an actual user/group name to the user/group numbers, we'll need passwd/group files to do so.  We'll need this to be placed on each partition, including system.  We already have world readable access to /{vendor,odm}/etc/{passwd,group}, but we'll need it for the rest of the partitions now too, hence this CL.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>