Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Lorenzo, thanks for the feedback. Left some more context.
> > 
> > Well, WifiNative.java (if it runs in the system server) definitely has permissions to do this.
> > 
> > Roshan: Ideally wifi code whould be responsible for creating the interfaces that it needs based on device configuration instead of some startup script running an "iw" command. how does this work on other devices, such as STA+AP devices? Who creates the interface in that case? Is it the wifi HAL?
> 
> For Pixel devices, we have worked with Qcom to ensure that all the available interfaces are created by the driver itself on load. We don't support creating interfaces in the default HAL implementation on AOSP. But, that implementation can be modified to dynamically create ifaces if needed for auto devices. There was a proposed CL from Qcom here: https://partner-android-review.googlesource.com/c/platform/hardware/qcom/wlan/+/1210190, but that is not fully fleshed out (need changes in the HIDL shim).

Lorenzo, updating WifiNative.java is definitely an option as you mentioned, however that would most likely be a new feature which may not be possible due to the associated bug's tags.

If we go with iw, it'll free us up to use it while not relying on drivers as Roshan mentioned. At the same time, other non-auto teams who are currently using iw won't need to put in device specific policies for it (as they are already doing).

Let me know your thoughts. I'm also happy to try other approaches that meet's the bug's commitments/tags. We just figured this would be the best approach for the reasons above.

Patch-set: 1
Reviewer: Gerrit User 1470463 <1470463@85c56323-6fa9-3386-8a01-6480fb634889>