Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > 
> > I really think we should avoid any solution that doesn't involve adding CAP_NET_* to new binaries. How about a hack in WifiNative in the short term, and a HAL-style solution in the longer term?
> 
> There isn't a simple fix in WifiNative (java framework) from what I know. Basically, this requires a duplication of the native code used by "iw" to create interfaces dynamically. We could do it in the HAL (native code), but again this requires duplication of the corresponding iw code.

Lorenzo and Roshan, thanks for all the feedback. 

Lorenzo, based off of Roshan's comment, it doesn't appear that updating WIfiNative in the short term will be an option. Happy to hear any other possible solutions?

If not and you feel your concerns addressed, can we go the iw approach as proposed? As iw is used to configure network interfaces, it does lend itself as a good candidate for these permissions. Also, if it helps, it looks as if net_raw is already used in quite a few places as per http://cs/search/?q=allow.*net_raw+package:%5Eandroid$&type=cs. 

I'm not suggesting that this in itself is reason to use it flippantly, just worth noting that it appears to be acceptable to use given the appropriate circumstances. Hope this helps!

Patch-set: 1
Reviewer: Gerrit User 1470463 <1470463@85c56323-6fa9-3386-8a01-6480fb634889>