Update patch set 2
Patch Set 2:

(1 comment)

Can you help me understand why this is implemented as a property, instead of just hard coding the selection into the code? Do we expect this to be device specific? Or do we expect that we'll be dynamically changing this value? Because of the sensitive nature of the linker, the linker seems like the last place where we want decisions about how the linker executes delegated to a property. Instead, it feels like everything the linker does should come be determined by code on the /system partition or any config files the linker itself generates.

Patch-set: 2
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>