Update patch set 2
Patch Set 2:

(1 comment)

> Original purpose of this work is 'make a way to let linker use /system/etc/ld.config.txt in case of linkerconfig generator does not work properly'. I could not be 100% sure that linkerconfig generator perfectly replaces existing ld.config.txt for now, so if I find anyone faces a problem with generated linker config, I thought giving a way to set a property would be better than give a patch for linker binary and let them rebuild it. I will remove this property once it looks like safe as it is. Do you think suggesting a property would be too much or unnecessary for this case?

Ok, so this is just temporary until you've proven that the linkerconfig generator is safe, and to give other people an easy way to switch back and forth to test the change? And we can revert this change after your testing is complete?

If so, I'm fine with this. The description made it sound like it was a longer term change.

One nit....

Patch-set: 2
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>