Update patch set 3
Patch Set 3: Code-Review+1

I think the dontaudit was the only way we figured this could work, because adbd doesn't know not to probe for a vsocket if it's enabled in the kernel but not set up by the host. And because the vsock_socket class is available in 4.14, and 4.14 is a supported kernel in pi, this does indeed seem like the right change to make, and the right place to make it.

Cody, could you confirm the above and +2 this CL if you agree?

Patch-set: 3
Reviewer: Gerrit User 1291337 <1291337@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1