Update patch set 1
Patch Set 1:

(I wasn't able to find this info in the bug and in the source tree) Hi, could you let us know where the sysprops are read and written?

public/property_contexts is a place for sysprops that are defined by platform (i.e. a module or a process in system.img) and are stable enough to be exposed to non-platform partitions like vendor.img.

Does this property satisfy above characteristics? If so, we need the Qualcomm change for surfaceflinger mentioned in the bug in AOSP.

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>