Update patch set 7
Patch Set 7:

Status: I'm currently waiting for Alan Stokes' response on the open comment threads from patchset 5.

If I get +2 on this CL then I intend to remove it from the CL topic and to submit it independently. This is because my understanding is that the other CLs in this topic depend on this one but not vice-versa, and I'd like to keep the CL topic for the final step as small as possible (for easy reverts). If you disagree then please protest.

Patch-set: 7
Reviewer: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>