Create change
Uploaded patch set 1.

Patch-set: 1
Change-id: If29cb22b2a7a5ce1b25d45ef8635e6cb81103327
Subject: This change is part of a topic that moves the recovery resources from the system partition to the vendor partition, if it exists, or the vendor directory on the system partition otherwise. The recovery resources are moving from the system image to the vendor partition so that a single system image may be used with either an A/B or a non-A/B vendor image. The topic removes a delta in the system image that prevented such reuse in the past.
Branch: refs/heads/master
Status: new
Topic: 
Commit: 6c949dca457143b7fe85316a5245228940a2e62a
Tag: autogenerated:gerrit:newPatchSet
Groups: 6c949dca457143b7fe85316a5245228940a2e62a
Private: false
Work-in-progress: false