Update patch set 3
Patch Set 3:

> Could you instead use PRODUCT_PRODUCT_PROPERTIES to set this prop from /product?
> And this way you don't need to make sepolicy changes.

I totally agree with you, technically it is possible to avoid changing system/sepolicy.

But in my opinion, this is rather a workaround than a solution. To me, it looks more natural if a vendor is allowed to set properties via the corresponding `/vendor` partition.

I am sure that you have more experience in SELinux, so if you confirm that this is not a bug but expected behavior I will abandon this CL.

Patch-set: 3
Reviewer: Gerrit User 1182749 <1182749@85c56323-6fa9-3386-8a01-6480fb634889>