Update patch set 1
Patch Set 1:

> Patch Set 1: Build-Cop-Override+1
> 
> stamping this on the basis that
> - it seems like it was just a mistake it was missed in r.android.com/1131912
> - reverting the internal topic looks a bit messy since it has cleanspec's that need to be undone etc
> 
> Jeff, please review this when you get a chance.

Just to add more context. The wifi stack has SAR functionality (limit wifi radio power when a call is in progress or next to the user's head/body). So, it needs to listen to audio events for triggering SAR. This is an existing functionality and the wifi stack was already doing this when it was inside system_server.

Patch-set: 1
Reviewer: Gerrit User 1074459 <1074459@85c56323-6fa9-3386-8a01-6480fb634889>