Update patch set 5
Patch Set 5:

-2 to hope to catch this before it auto submits.

Why do we have 'vendor' in the name of the property?  That namespace is reserved for vendors to create their own properties without conflicting against Google created ones (like this).  There are a handful of exceptions like the vendor.build properties below, but if this is a Google defined property that we just allow vendors to set, then it shouldn't be in the vendor. namespace.

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>