Update patch set 1
Patch Set 1:

Should this be set by the vendor partition or the product image?  I know that the email thread we had mentioned the vendor partition, but I'm thinking maybe this should be a 'product' decision?

Or do we need to put sanitizer libraries in the vendor partition anyway, so we're already tied to that?

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>