Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think this is agnostic to the partition. This is determined by the way build command was invoked. So, ideally, we need this sysprop for all partitions, but practically we need this only for the system partition.

The key thing that makes a build a "hwasan" build is the presence of -fsanitize=hwaddress when building libc. So it's tied to the system image.

In that email thread you mention, we need it visible in a vendor .rc file. There is no reason to restrict visibility at all, think something like "ro.build.id".

Patch-set: 1