Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> IIRC, most device launched since Oreo has a /vendor partition (for Treble compliance).
> Non-Full-Treble devices might not be well supported in AOSP now (master refers to R currently).
> i.e., there might be something else will break non-full-treble devices.
> 
> Is this for a device launched before Oreo?

can we get an answer to this question please?

In the interest of keeping things simple and eventually moving away from /system/vendor, I would prefer if we not add more support for this file path. The goal is for each file / directory to have 1 and exactly 1 location, and this change moves away from that ideal. I'd like to get a better understanding of the "one true path" and see if we can somehow avoid this change.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>