Update patch set 1
Patch Set 1:

> Patch Set 1: Verified+1 Code-Review+1
> 
> The sword of the righteous shall strike Bowgo and Siarhei... https://android.googlesource.com/platform/system/sepolicy/+/3639f5796 & https://android.googlesource.com/platform/system/sepolicy/+/591293111
> 

Hi,
Just to be clear, the above two example are fallback to /vendor/odm, when there is no /odm partition. (it assumes there is a /vendor partition). So it's not quite related to the fallback of /system/vendor here.

Note that I'm not oppose to the change here, just wanted to get a better understanding of the use case. Also, there might be something else that breaks non-full-treble devices.
It will be great if all related CLs can be uploaded together with the same 'topic' (or at least group them by using 'hashtag').

Just wanted to prevent the surprise that fixing this still doesn't work on non-full-treble devices.


> Kidding aside, legacy (I hate that word) devices are still around and need this.

Patch-set: 1
Reviewer: Gerrit User 1080813 <1080813@85c56323-6fa9-3386-8a01-6480fb634889>