Update patch set 1
Patch Set 1:

We had to revert the change to ART because it broke the bots. 

We've had a revert^2 for that all day but have been hampered by pre-submit issues (http://b/142719132).

In the interim, we should revert the sepolicy changes. We'll redo them with the art revert^2 change.

Patch-set: 1