Update patch set 1
Patch Set 1:

I don't think we can/should do this in a Treble world.  When we made the original property exported, we signed it up to be a backwards compatible API.  What would happen if an old vendor image set the old property and the system image did not read it?

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>